# docs/api/clubspot.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ClubSpot.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/people": {
      "get": {
        "tags": [
          "people"
        ],
        "operationId": "SearchPeople",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PeoplePageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "people"
        ],
        "operationId": "CreatePerson",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePersonRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/people/{id}": {
      "get": {
        "tags": [
          "people"
        ],
        "operationId": "GetPerson",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonDetailsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/people/blocks": {
      "post": {
        "tags": [
          "people"
        ],
        "operationId": "BlockPeople",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockPeopleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AffectedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/people/{id}/block": {
      "put": {
        "tags": [
          "people"
        ],
        "operationId": "SetPersonBlock",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetBlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/people/{id}/notes": {
      "post": {
        "tags": [
          "people"
        ],
        "operationId": "AddPersonNote",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/people/{id}/payments": {
      "post": {
        "tags": [
          "people"
        ],
        "operationId": "RegisterPersonPayment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AddNoteRequest": {
        "required": [
          "text"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          }
        }
      },
      "AffectedResponse": {
        "required": [
          "affected"
        ],
        "type": "object",
        "properties": {
          "affected": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "BlockPeopleRequest": {
        "required": [
          "ids",
          "blocked"
        ],
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "blocked": {
            "type": "boolean"
          }
        }
      },
      "BlockResponse": {
        "required": [
          "blocked"
        ],
        "type": "object",
        "properties": {
          "blocked": {
            "type": "boolean"
          }
        }
      },
...
      "CreatePersonRequest": {
        "required": [
          "name",
          "phone",
          "email"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
...
      "NoteResponse": {
        "required": [
          "text",
          "authorName",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PaymentResponse": {
        "required": [
          "paid"
        ],
        "type": "object",
        "properties": {
          "paid": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "PeoplePageResponse": {
        "required": [
          "items",
          "total",
          "page",
          "pages",
          "pageSize",
          "census",
          "needsAttention",
          "totalDebt",
          "totals"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PersonResponse"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pages": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "census": {
            "type": "integer",
            "format": "int32"
          },
          "needsAttention": {
            "type": "integer",
            "format": "int32"
          },
          "totalDebt": {
            "type": "number",
            "format": "double"
          },
          "totals": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
...
      "PersonDetailsResponse": {
        "required": [
          "person",
          "notes"
        ],
        "type": "object",
        "properties": {
          "person": {
            "$ref": "#/components/schemas/PersonResponse"
          },
          "notes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoteResponse"
            }
          }
        }
      },
      "PersonOrigin": {
        "enum": [
          "app",
          "counter"
        ],
        "type": "string"
      },
      "PersonResponse": {
        "required": [
          "id",
          "name",
          "phone",
          "email",
          "origin",
          "bookings",
          "lastBookingOn",
          "debt",
          "isBlocked",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "origin": {
            "$ref": "#/components/schemas/PersonOrigin"
          },
          "bookings": {
            "type": "integer",
            "format": "int32"
          },
          "lastBookingOn": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "debt": {
            "type": "number",
            "format": "double"
          },
          "isBlocked": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SetBlockRequest": {
        "required": [
          "blocked"
        ],
        "type": "object",
        "properties": {
          "blocked": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}